tests: Move transformClassicToBidiSelector per review comment
